#415 get rid of fiscal_bricks_to_fiscal_inx
