--- src/f10_etl/test/test_transformer_event_bricks_to_fiscal_bricks.py
@@ -1,165 +1,0 @@
-from src.f00_instrument.file import create_path, set_dir, delete_dir
-from src.f04_gift.atom_config import face_id_str, fiscal_id_str
-from src.f07_fiscal.fiscal_config import cumlative_minute_str, hour_label_str
-from src.f08_pidgin.pidgin_config import event_id_str
-from src.f09_brick.pandas_tool import upsert_sheet, forge_valid_str, sheet_exists
-from src.f10_etl.transformers import (
-    etl_event_bricks_to_fiscal_bricks,
-    get_fiscal_events_by_dirs,
-)
-from src.f10_etl.examples.etl_env import get_test_etl_dir, env_dir_setup_cleanup
-from pandas.testing import (
-    assert_frame_equal as pandas_assert_frame_equal,
-)
-from pandas import DataFrame, read_excel as pandas_read_excel
-from os.path import exists as os_path_exists
-
-
-def test_etl_event_bricks_to_fiscal_bricks_CreatesFaceBrickSheets_Scenario0_MultpleFaceIDs(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    sue_str = "Sue"
-    zia_str = "Zia"
-    event3 = 3
-    event7 = 7
-    event9 = 9
-    minute_360 = 360
-    minute_420 = 420
-    hour6am = "6am"
-    hour7am = "7am"
-    br00003_columns = [
-        face_id_str(),
-        event_id_str(),
-        fiscal_id_str(),
-        hour_label_str(),
-        cumlative_minute_str(),
-    ]
-    music23_str = "music23"
-    music55_str = "music55"
-    sue0 = [sue_str, event3, music23_str, hour6am, minute_360]
-    sue1 = [sue_str, event3, music23_str, hour7am, minute_420]
-    zia0 = [zia_str, event7, music23_str, hour7am, minute_420]
-    zia1 = [zia_str, event9, music23_str, hour6am, minute_360]
-    zia2 = [zia_str, event9, music23_str, hour7am, minute_420]
-    zia3 = [zia_str, event9, music55_str, hour7am, minute_420]
-    example_event3_df = DataFrame([sue0, sue1], columns=br00003_columns)
-    example_event7_df = DataFrame([zia0], columns=br00003_columns)
-    example_event9_df = DataFrame([zia1, zia2, zia3], columns=br00003_columns)
-    x_etl_dir = get_test_etl_dir()
-    x_faces_otx_dir = create_path(x_etl_dir, "faces_otx")
-    br00003_filename = "br00003.xlsx"
-    sue_dir = create_path(x_faces_otx_dir, sue_str)
-    zia_dir = create_path(x_faces_otx_dir, zia_str)
-    event3_dir = create_path(sue_dir, event3)
-    event7_dir = create_path(zia_dir, event7)
-    event9_dir = create_path(zia_dir, event9)
-    event3_br00003_filepath = create_path(event3_dir, br00003_filename)
-    event7_br00003_filepath = create_path(event7_dir, br00003_filename)
-    event9_br00003_filepath = create_path(event9_dir, br00003_filename)
-    upsert_sheet(event3_br00003_filepath, forge_valid_str(), example_event3_df)
-    upsert_sheet(event7_br00003_filepath, forge_valid_str(), example_event7_df)
-    upsert_sheet(event9_br00003_filepath, forge_valid_str(), example_event9_df)
-    e3_music23_dir = create_path(event3_dir, music23_str)
-    e7_music23_dir = create_path(event7_dir, music23_str)
-    e9_music23_dir = create_path(event9_dir, music23_str)
-    e9_music55_dir = create_path(event9_dir, music55_str)
-    e3_music23_br00003_filepath = create_path(e3_music23_dir, br00003_filename)
-    e7_music23_br00003_filepath = create_path(e7_music23_dir, br00003_filename)
-    e9_music23_br00003_filepath = create_path(e9_music23_dir, br00003_filename)
-    e9_music55_br00003_filepath = create_path(e9_music55_dir, br00003_filename)
-    print(f"{e3_music23_br00003_filepath=}")
-    print(f"{e7_music23_br00003_filepath=}")
-    print(f"{e9_music23_br00003_filepath=}")
-    print(f"{e9_music55_br00003_filepath=}")
-    assert sheet_exists(e3_music23_br00003_filepath, forge_valid_str()) is False
-    assert sheet_exists(e7_music23_br00003_filepath, forge_valid_str()) is False
-    assert sheet_exists(e9_music23_br00003_filepath, forge_valid_str()) is False
-    assert sheet_exists(e9_music55_br00003_filepath, forge_valid_str()) is False
-
-    # WHEN
-    etl_event_bricks_to_fiscal_bricks(x_faces_otx_dir)
-
-    # THEN
-    assert sheet_exists(e7_music23_br00003_filepath, forge_valid_str())
-    assert sheet_exists(e3_music23_br00003_filepath, forge_valid_str())
-    assert sheet_exists(e9_music23_br00003_filepath, forge_valid_str())
-    assert sheet_exists(e9_music55_br00003_filepath, forge_valid_str())
-    gen_e3_music23_df = pandas_read_excel(
-        e3_music23_br00003_filepath, forge_valid_str()
-    )
-    gen_e7_music23_df = pandas_read_excel(
-        e7_music23_br00003_filepath, forge_valid_str()
-    )
-    gen_e9_music23_df = pandas_read_excel(
-        e9_music23_br00003_filepath, forge_valid_str()
-    )
-    gen_e9_music55_df = pandas_read_excel(
-        e9_music55_br00003_filepath, forge_valid_str()
-    )
-    expected_e9_music23_df = DataFrame([zia1, zia2], columns=br00003_columns)
-    expected_e9_music55_df = DataFrame([zia3], columns=br00003_columns)
-    print("gen_e9_music55_df")
-    print(f"{gen_e9_music55_df}")
-    print("expected_e9_music55_df")
-    print(f"{expected_e9_music55_df}")
-    pandas_assert_frame_equal(gen_e3_music23_df, example_event3_df)
-    pandas_assert_frame_equal(gen_e7_music23_df, example_event7_df)
-    pandas_assert_frame_equal(gen_e9_music23_df, expected_e9_music23_df)
-    pandas_assert_frame_equal(gen_e9_music55_df, expected_e9_music55_df)
-
-
-def test_get_fiscal_events_by_dirs_SetsAttr(env_dir_setup_cleanup):
-    # ESTABLISH
-    sue_str = "Sue"
-    zia_str = "Zia"
-    event3 = 3
-    event7 = 7
-    event9 = 9
-    music23_str = "music23"
-    music55_str = "music55"
-    x_etl_dir = get_test_etl_dir()
-    x_faces_otx_dir = create_path(x_etl_dir, "faces_otx")
-    sue_dir = create_path(x_faces_otx_dir, sue_str)
-    zia_dir = create_path(x_faces_otx_dir, zia_str)
-    event3_dir = create_path(sue_dir, event3)
-    event7_dir = create_path(zia_dir, event7)
-    event9_dir = create_path(zia_dir, event9)
-    e3_music23_dir = create_path(event3_dir, music23_str)
-    e7_music23_dir = create_path(event7_dir, music23_str)
-    e9_music23_dir = create_path(event9_dir, music23_str)
-    e9_music55_dir = create_path(event9_dir, music55_str)
-    set_dir(e3_music23_dir)
-    set_dir(e7_music23_dir)
-    set_dir(e9_music23_dir)
-    set_dir(e9_music55_dir)
-    print(f"{e3_music23_dir=}")
-    print(f"{e7_music23_dir=}")
-    print(f"{e9_music23_dir=}")
-    print(f"{e9_music55_dir=}")
-    assert os_path_exists(e3_music23_dir)
-    assert os_path_exists(e7_music23_dir)
-    assert os_path_exists(e9_music23_dir)
-    assert os_path_exists(e9_music55_dir)
-
-    # WHEN
-    x_fiscal_events = get_fiscal_events_by_dirs(x_faces_otx_dir)
-
-    # THEN
-    assert x_fiscal_events == {
-        music23_str: {event3, event7, event9},
-        music55_str: {event9},
-    }
-
-    # WHEN
-    delete_dir(e9_music55_dir)
-    assert os_path_exists(e3_music23_dir)
-    assert os_path_exists(e7_music23_dir)
-    assert os_path_exists(e9_music23_dir)
-    assert os_path_exists(e9_music55_dir) is False
-
-    # WHEN
-    x_fiscal_events = get_fiscal_events_by_dirs(x_faces_otx_dir)
-
-    # THEN
-    assert x_fiscal_events == {music23_str: {event3, event7, event9}}
