+++ src/f11_world/world.py
@@ -30,7 +30,6 @@ from src.f10_etl.transformers import (
     etl_pidgin_jsons_inherit_younger_pidgins,
     etl_forge_bricks_to_otx_face_bricks,
     etl_face_bricks_to_event_bricks,
-    etl_event_bricks_to_fiscal_bricks,
     get_fiscal_events_by_dirs,
     get_pidgin_events_by_dirs,
 )
@@ -60,8 +59,6 @@ class WorldUnit:
     _fiscalunits: set[FiscalID] = None
     _fiscal_events: dict[FiscalID, set[TimeLinePoint]] = None
     _pidgin_events: dict[FaceID, set[TimeLinePoint]] = None
-    # _fiscal_pidgins: dict fiscal_id: dict [fiscal_event_id, associated_pidgin_event_id]
-    _fiscal_pidgins: dict[FiscalID, dict[TimeLinePoint, TimeLinePoint]] = None
 
     def set_event(self, event_id: TimeLinePoint, face_id: FaceID):
         self.events[event_id] = face_id
@@ -84,37 +81,6 @@ class WorldUnit:
 
     def _set_pidgin_events(self):
         self._pidgin_events = get_pidgin_events_by_dirs(self._faces_otx_dir)
-
-    def _set_fiscal_pidgins(self):
-        self._fiscal_pidgins = {}
-        for fiscal_id, fiscal_events in self._fiscal_events.items():
-            self._fiscal_pidgins[fiscal_id] = {}
-            for fiscal_event_id in fiscal_events:
-                self._set_fiscal_pidgin_max_prev_pidgin_event_id(
-                    fiscal_id, fiscal_event_id
-                )
-
-    def _set_fiscal_pidgin_max_prev_pidgin_event_id(
-        self, fiscal_id: FiscalID, fiscal_event_id: TimeLinePoint
-    ):
-        fiscal_face_id = self.events.get(fiscal_event_id)
-        if fiscal_face_id is None:
-            exception_str = (
-                f"fiscal_event_id {fiscal_event_id} does not have associated face_id"
-            )
-            raise _set_fiscal_pidgin_Exception(exception_str)
-        face_pidgin_event_ids = self._pidgin_events.get(fiscal_face_id)
-        face_pidgin_event_ids = get_empty_set_if_None(face_pidgin_event_ids)
-        max_prev_pidgin_event_id = max(
-            (
-                pidgin_event_id
-                for pidgin_event_id in face_pidgin_event_ids
-                if pidgin_event_id <= fiscal_event_id
-            ),
-            default=None,
-        )
-        x_pidgins = self._fiscal_pidgins[fiscal_id]
-        x_pidgins[fiscal_event_id] = max_prev_pidgin_event_id
 
     def set_mine_dir(self, x_dir: str):
         self._mine_dir = x_dir
@@ -184,11 +150,6 @@ class WorldUnit:
     def face_bricks_to_event_bricks(self):
         etl_face_bricks_to_event_bricks(self._faces_otx_dir)
 
-    def event_bricks_to_fiscal_bricks(self):
-        etl_event_bricks_to_fiscal_bricks(self._faces_otx_dir)
-        self._set_fiscal_events()
-        self._set_fiscal_pidgins()
-
     def get_dict(self) -> dict:
         return {
             "world_id": self.world_id,
@@ -220,7 +181,6 @@ def worldunit_shop(
         _mine_dir=mine_dir,
         _fiscal_events={},
         _pidgin_events={},
-        _fiscal_pidgins={},
     )
     x_worldunit._set_world_dirs()
     if not x_worldunit._mine_dir:
